+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 
 ### Notable changes
 
+* added predefined piece colors - library entries now can have pre-set default colors
+* added color support and "?", "!" and "↑" sides to the generic A-Z piece
 * added hotkeys F/R to flip/rotate backwards
 
 ### Other changes